refactor jaxrs-mysql-quarkus #392
