+++ application/jaxrs-mysql-quarkus/src/main/java/webapp/tier/healthcheck/LivenessHealthCheckMysql.java
@@ -8,7 +8,7 @@ import org.eclipse.microprofile.health.Liveness;
 
 @Liveness
 @ApplicationScoped
-public class LivenessHealthCheck implements HealthCheck {
+public class LivenessHealthCheckMysql implements HealthCheck {
 
 	@Override
 	public HealthCheckResponse call() {
