--- application/jaxrs-mysql-quarkus/src/main/java/webapp/tier/healthcheck/ReadinessHealthCheck.java
@@ -1,33 +1,0 @@
-package webapp.tier.healthcheck;
-
-import java.util.logging.Logger;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-
-import org.eclipse.microprofile.health.HealthCheck;
-import org.eclipse.microprofile.health.HealthCheckResponse;
-import org.eclipse.microprofile.health.Readiness;
-
-import webapp.tier.service.MysqlService;
-
-@Readiness
-@ApplicationScoped
-public class ReadinessHealthCheck implements HealthCheck {
-
-	private static final Logger LOG = Logger.getLogger(ReadinessHealthCheck.class.getSimpleName());
-
-	@Inject
-	MysqlService mysqlsvc;
-
-	@Override
-	public HealthCheckResponse call() {
-		if (mysqlsvc.connectionStatus()) {
-			LOG.fine("Readiness: UP");
-			return HealthCheckResponse.up("Database connection health check");
-		} else {
-			LOG.warning("Readiness: DOWN");
-			return HealthCheckResponse.down("Database connection health check");
-		}
-	}
-}
